chore: move files utils to separate file
